async updates on graph to reduce jitter
converted all code to doubles

git-svn-id: svn+ssh://ares.lids.mit.edu/code/pratikac@1653 d6d04add-692a-0410-8435-c93ef0a91bfa

+++ pf/kushner/systems/dubins.cpp
@@ -2,11 +2,11 @@
 
 System::System()
 {
-    min_states = new float[NUM_DIM];
-    max_states = new float[NUM_DIM];
-    obs_noise = new float[NUM_DIM];
-    process_noise = new float[NUM_DIM];
-    init_var = new float[NUM_DIM];
+    min_states = new double[NUM_DIM];
+    max_states = new double[NUM_DIM];
+    obs_noise = new double[NUM_DIM];
+    process_noise = new double[NUM_DIM];
+    init_var = new double[NUM_DIM];
 
     for(int i=0; i< NUM_DIM; i++)
     {
@@ -96,13 +96,13 @@ int System::get_key(State& s, double *key)
     return 0;
 }
 
-State System::integrate(State& s, float duration, bool is_clean)
+State System::integrate(State& s, double duration, bool is_clean)
 {
     State t;
 
-    float *var = new float[NUM_DIM];
-    float *mean = new float[NUM_DIM];
-    float *tmp = new float[NUM_DIM];
+    double *var = new double[NUM_DIM];
+    double *mean = new double[NUM_DIM];
+    double *tmp = new double[NUM_DIM];
 
     for(int i=0; i<NUM_DIM; i++)
     {
@@ -116,7 +116,7 @@ State System::integrate(State& s, float duration, bool is_clean)
         mean[i] = 0;
     }
         
-    float curr_time = 0;
+    double curr_time = 0;
     while(curr_time < duration)
     {
         if( !is_clean)  
@@ -136,7 +136,7 @@ State System::integrate(State& s, float duration, bool is_clean)
     return t;
 }
 
-void System::get_variance(State& s, float duration, float* var)
+void System::get_variance(State& s, double duration, double* var)
 {
     for(int i=0; i<NUM_DIM; i++)
     {   
@@ -149,8 +149,8 @@ State System::observation(State& s, bool is_clean)
 {
     State t;
 
-    float *tmp = new float[NUM_DIM];
-    float *mean = new float[NUM_DIM];
+    double *tmp = new double[NUM_DIM];
+    double *mean = new double[NUM_DIM];
 
     if( !is_clean)  
         multivar_normal( mean, obs_noise, tmp, NUM_DIM);
@@ -170,7 +170,7 @@ State System::observation(State& s, bool is_clean)
     return t;
 }
 
-void System::get_kalman_path( vector<State>& obs, vector<float>& obs_times, list<State>& kalman_path)
+void System::get_kalman_path( vector<State>& obs, vector<double>& obs_times, list<State>& kalman_path)
 {
     ;
 }
