+++ pf/kushner/systems/ship.cpp
@@ -2,11 +2,11 @@
 
 System::System()
 {
-    min_states = new float[NUM_DIM];
-    max_states = new float[NUM_DIM];
-    obs_noise = new float[NUM_DIM];
-    process_noise = new float[NUM_DIM];
-    init_var = new float[NUM_DIM];
+    min_states = new double[NUM_DIM];
+    max_states = new double[NUM_DIM];
+    obs_noise = new double[NUM_DIM];
+    process_noise = new double[NUM_DIM];
+    init_var = new double[NUM_DIM];
 
     for(int i=0; i< NUM_DIM; i++)
     {
@@ -67,13 +67,13 @@ int System::get_key(State& s, double *key)
     return 0;
 }
 
-State System::integrate(State& s, float duration, bool is_clean)
+State System::integrate(State& s, double duration, bool is_clean)
 {
     State t;
 
-    float *var = new float[NUM_DIM];
-    float *mean = new float[NUM_DIM];
-    float *tmp = new float[NUM_DIM];
+    double *var = new double[NUM_DIM];
+    double *mean = new double[NUM_DIM];
+    double *tmp = new double[NUM_DIM];
 
     for(int i=0; i<NUM_DIM; i++)
     {
@@ -87,15 +87,15 @@ State System::integrate(State& s, float duration, bool is_clean)
         mean[i] = 0;
     }
 
-    float curr_time = 0;
+    double curr_time = 0;
     while(curr_time < duration)
     {
         if( !is_clean)
             multivar_normal( mean, var, tmp, NUM_DIM);
         
-        float old_x3 = t.x[2]; float old_x4 = t.x[3];
+        double old_x3 = t.x[2]; double old_x4 = t.x[3];
         
-        float f1dt=0, f2dt = 0;
+        double f1dt=0, f2dt = 0;
         if( ( sqrt(t.x[0]*t.x[0] + t.x[1]*t.x[1]) >= 9) && ( (t.x[0]*t.x[2] + t.x[1]*t.x[3]) >= 0) )
         {
             f1dt = -50*sim_time_delta*t.x[0]/sqrt(t.x[0]*t.x[0] + t.x[1]*t.x[1]);
@@ -117,7 +117,7 @@ State System::integrate(State& s, float duration, bool is_clean)
     return t;
 }
 
-void System::get_variance(State& s, float duration, float* var)
+void System::get_variance(State& s, double duration, double* var)
 {
     for(int i=0; i<NUM_DIM; i++)
     {   
@@ -129,8 +129,8 @@ State System::observation(State& s, bool is_clean)
 {
     State t;
 
-    float *tmp = new float[NUM_DIM_OBS];
-    float *mean = new float[NUM_DIM_OBS];
+    double *tmp = new double[NUM_DIM_OBS];
+    double *mean = new double[NUM_DIM_OBS];
 
     if( !is_clean)  
         multivar_normal( mean, obs_noise, tmp, NUM_DIM_OBS);
@@ -153,7 +153,7 @@ State System::observation(State& s, bool is_clean)
     return t;
 }
 
-void System::get_kalman_path( vector<State>& obs, vector<float>& obs_times, list<State>& kalman_path)
+void System::get_kalman_path( vector<State>& obs, vector<double>& obs_times, list<State>& kalman_path)
 {
     return;
 }
